[GIE Compiler] introduce ir meta fetching api
