+++ interactive_engine/compiler/conf/ir.compiler.properties
@@ -61,15 +61,3 @@ calcite.default.charset: UTF-8
 
 # set the max capacity of the result streaming buffer for each query
 # per.query.stream.buffer.max.capacity: 256
-
-# set the mode to read ir meta, either from a 'local' file or via remote 'http'.
-# ir.meta.reader.mode: local
-
-# if ir.meta.reader.mode is 'http', set the remote ir meta service host
-# ir.meta.service.host: localhost:8080
-
-# if ir.meta.reader.mode is 'http', set the interval in milliseconds to fetch ir meta
-# ir.meta.fetch.interval.ms: 1000
-
-# if ir.meta.reader.mode is 'http', set the timeout in milliseconds to fetch ir graph statistics
-# ir.statistics.fetch.interval.ms: 86400000l
