--- interactive_engine/compiler/src/main/java/com/alibaba/graphscope/common/client/channel/MetaServiceChannelFetcher.java
@@ -1,47 +1,0 @@
-/*
- *
- *  * Copyright 2020 Alibaba Group Holding Limited.
- *  *
- *  * Licensed under the Apache License, Version 2.0 (the "License");
- *  * you may not use this file except in compliance with the License.
- *  * You may obtain a copy of the License at
- *  *
- *  * http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
- *
- */
-
-package com.alibaba.graphscope.common.client.channel;
-
-import com.alibaba.graphscope.common.config.Configs;
-import com.alibaba.graphscope.common.config.FrontendConfig;
-import com.google.common.collect.ImmutableList;
-
-import java.net.URI;
-import java.util.List;
-
-// read meta service host from config and return it as URI format
-public class MetaServiceChannelFetcher implements ChannelFetcher<URI> {
-    private static final String schema = "http";
-    private final Configs graphConfig;
-
-    public MetaServiceChannelFetcher(Configs graphConfig) {
-        this.graphConfig = graphConfig;
-    }
-
-    @Override
-    public List<URI> fetch() {
-        String host = FrontendConfig.IR_META_SERVICE_HOST.get(graphConfig);
-        return ImmutableList.of(URI.create(schema + "://" + host));
-    }
-
-    @Override
-    public Type getType() {
-        return Type.HTTP;
-    }
-}
