--- interactive_engine/compiler/src/main/java/com/alibaba/graphscope/common/ir/meta/fetcher/DynamicIrMetaFetcher.java
@@ -1,108 +1,0 @@
-/*
- *
- *  * Copyright 2020 Alibaba Group Holding Limited.
- *  *
- *  * Licensed under the Apache License, Version 2.0 (the "License");
- *  * you may not use this file except in compliance with the License.
- *  * You may obtain a copy of the License at
- *  *
- *  * http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
- *
- */
-
-package com.alibaba.graphscope.common.ir.meta.fetcher;
-
-import com.alibaba.graphscope.common.config.Configs;
-import com.alibaba.graphscope.common.config.FrontendConfig;
-import com.alibaba.graphscope.common.ir.meta.IrMeta;
-import com.alibaba.graphscope.common.ir.meta.IrMetaStats;
-import com.alibaba.graphscope.common.ir.meta.IrMetaTracker;
-import com.alibaba.graphscope.common.ir.meta.reader.IrMetaReader;
-import com.alibaba.graphscope.groot.common.schema.api.GraphStatistics;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.Optional;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ScheduledThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
-
-/**
- * Periodically update IrMeta, with the update frequency controlled by configuration.
- * Specifically, for procedures, a remote update will be actively triggered when they are not found locally.
- */
-public class DynamicIrMetaFetcher extends IrMetaFetcher implements AutoCloseable {
-    private static final Logger logger = LoggerFactory.getLogger(DynamicIrMetaFetcher.class);
-    private final ScheduledExecutorService scheduler;
-    private volatile IrMetaStats currentState;
-
-    public DynamicIrMetaFetcher(Configs configs, IrMetaReader dataReader, IrMetaTracker tracker) {
-        super(dataReader, tracker);
-        this.scheduler = new ScheduledThreadPoolExecutor(2);
-        this.scheduler.scheduleAtFixedRate(
-                () -> syncMeta(),
-                0,
-                FrontendConfig.IR_META_FETCH_INTERVAL_MS.get(configs),
-                TimeUnit.MILLISECONDS);
-        this.scheduler.scheduleAtFixedRate(
-                () -> syncStats(),
-                0,
-                FrontendConfig.IR_STATISTICS_FETCH_INTERVAL_MS.get(configs),
-                TimeUnit.MILLISECONDS);
-    }
-
-    @Override
-    public Optional<IrMeta> fetch() {
-        return currentState == null ? Optional.empty() : Optional.of(currentState);
-    }
-
-    private synchronized void syncMeta() {
-        try {
-            IrMeta meta = this.reader.readMeta();
-            this.currentState =
-                    new IrMetaStats(
-                            meta.getSnapshotId(),
-                            meta.getSchema(),
-                            meta.getStoredProcedures(),
-                            this.currentState == null ? null : this.currentState.getStatistics());
-            if (this.currentState.getStatistics() == null) {
-                syncStats();
-            }
-        } catch (Exception e) {
-            logger.warn("failed to read meta data, error is {}", e);
-        }
-    }
-
-    private synchronized void syncStats() {
-        try {
-            if (this.currentState != null) {
-                GraphStatistics stats = this.reader.readStats(this.currentState.getGraphId());
-                if (stats != null) {
-                    this.currentState =
-                            new IrMetaStats(
-                                    this.currentState.getSnapshotId(),
-                                    this.currentState.getSchema(),
-                                    this.currentState.getStoredProcedures(),
-                                    stats);
-                    if (tracker != null) {
-                        tracker.onChanged(this.currentState);
-                    }
-                }
-            }
-        } catch (Exception e) {
-            logger.warn("failed to read graph statistics, error is {}", e);
-        }
-    }
-
-    @Override
-    public void close() throws Exception {
-        this.scheduler.shutdown();
-    }
-}
