--- interactive_engine/compiler/src/main/java/com/alibaba/graphscope/common/ir/meta/reader/SchemaInputStream.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2020 Alibaba Group Holding Limited.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.graphscope.common.ir.meta.reader;
-
-import java.io.InputStream;
-
-public class SchemaInputStream {
-    private final InputStream inputStream;
-    private final FileFormatType formatType;
-
-    public SchemaInputStream(InputStream inputStream, FileFormatType formatType) {
-        this.inputStream = inputStream;
-        this.formatType = formatType;
-    }
-
-    public InputStream getInputStream() {
-        return inputStream;
-    }
-
-    public FileFormatType getFormatType() {
-        return formatType;
-    }
-}
