+++ interactive_engine/groot-server/src/main/java/com/alibaba/graphscope/groot/servers/ir/IrServiceProducer.java
@@ -22,7 +22,6 @@ import com.alibaba.graphscope.common.config.AuthConfig;
 import com.alibaba.graphscope.common.config.FrontendConfig;
 import com.alibaba.graphscope.common.config.PegasusConfig;
 import com.alibaba.graphscope.common.config.PlannerConfig;
-import com.alibaba.graphscope.common.ir.meta.fetcher.DynamicIrMetaFetcher;
 import com.alibaba.graphscope.common.ir.meta.fetcher.IrMetaFetcher;
 import com.alibaba.graphscope.common.ir.planner.GraphRelOptimizer;
 import com.alibaba.graphscope.gremlin.integration.result.TestGraphFactory;
@@ -61,10 +60,8 @@ public class IrServiceProducer {
         logger.info("IR configs: {}", irConfigs);
         GraphRelOptimizer optimizer = new GraphRelOptimizer(irConfigs);
         IrMetaFetcher irMetaFetcher =
-                new DynamicIrMetaFetcher(
-                        irConfigs,
-                        new GrootIrMetaReader(schemaFetcher),
-                        optimizer.getGlogueHolder());
+                new GrootMetaFetcher(
+                        new GrootIrMetaReader(schemaFetcher), optimizer.getGlogueHolder());
         RoleClients<SnapshotUpdateClient> updateCommitter =
                 new RoleClients<>(channelManager, RoleType.COORDINATOR, SnapshotUpdateClient::new);
         int frontendId = CommonConfig.NODE_IDX.get(configs);
@@ -133,10 +130,6 @@ public class IrServiceProducer {
         addToConfigMapIfExist(FrontendConfig.FRONTEND_SERVER_NUM.getKey(), configMap);
         // add frontend qps limit
         addToConfigMapIfExist(FrontendConfig.QUERY_PER_SECOND_LIMIT.getKey(), configMap);
-        // add ir meta fetch interval
-        addToConfigMapIfExist(FrontendConfig.IR_META_FETCH_INTERVAL_MS.getKey(), configMap);
-        // add ir statistics fetch interval
-        addToConfigMapIfExist(FrontendConfig.IR_STATISTICS_FETCH_INTERVAL_MS.getKey(), configMap);
         // add graph planner configs
         addToConfigMapIfExist(PlannerConfig.GRAPH_PLANNER_IS_ON.getKey(), configMap);
         addToConfigMapIfExist(PlannerConfig.GRAPH_PLANNER_OPT.getKey(), configMap);
